refactor : order checkstyle 적용
