--- pickly-service/src/main/java/org/pickly/service/member/notification/service/interfaces/NotificationStandardService.java
@@ -1,9 +1,0 @@
-package org.pickly.service.member.notification.service.interfaces;
-
-import org.pickly.service.member.notification.entity.NotificationStandard;
-
-public interface NotificationStandardService {
-
-  NotificationStandard findByMember(Long memberId);
-
-}
