+++ pickly-service/src/main/java/org/pickly/service/notification/service/interfaces/NotificationStandardService.java
@@ -1,0 +1,9 @@
+package org.pickly.service.notification.service.interfaces;
+
+import org.pickly.service.notification.entity.NotificationStandard;
+
+public interface NotificationStandardService {
+
+  NotificationStandard findByMember(Long memberId);
+
+}
